moved non user-facing objects and operations to internals.py
