+++ chirho/dynamical/ops/__init__.py
@@ -1,0 +1,1 @@
+from .dynamical import Dynamics, simulate
